Added synthetic data-fitting NLME
Inside Matlab (in the correct directory), call "synthetic_NLME(num_expts,num_params)", where num_expts is the number of experiments/datasets, and num_params is the number of parameters being fit; 2 if fitting Hill and pIC50 and 1 if only fitting pIC50, where Hill is assumed to be 1.

In the latter case, Hill=1 is also used to generate the synthetic data, as well as fitting pIC50. I'm not currently sure if we should use a slightly different value for Hill when generating the synthetic data, and then still keep Hill=1 when fitting pIC50.

Can play with the "true" fixed effect values and the "true" random effects variances and the observation noise sigma. Can also change the number of concentrations, i.e. number of data points per experiment/dataset.

When generating the data, it generates some parameter values from the mixed model, solves the dose-response model, then adds Normal noise with standard deviation sigma. It then truncates anything less than 0 or greater than